+++ opendbc/safety/safety_helpers.h
@@ -25,3 +25,12 @@ static float interpolate(struct lookup_t xy, float x) {
   }
   return ret;
 }
+
+bool max_limit_check(int val, const int MAX_VAL, const int MIN_VAL) {
+  return (val > MAX_VAL) || (val < MIN_VAL);
+}
+
+int ROUND(float val) {
+  return val + ((val > 0.0) ? 0.5 : -0.5);
+}
+
